requirements.txt
# === DocuMind v3 — Production Dependencies ===
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic-settings==2.4.0
python-dotenv==1.0.1

# PDF Processing
pymupdf==1.25.3

# ML / Embeddings / Search
numpy>=1.24.0
scikit-learn>=1.3.0

# HTTP Client
httpx>=0.24.1

# Async SSE Streaming
sse-starlette>=1.6.5

# BM25 Search
rank-bm25>=0.2.2
